Increases speed of doing a run by adding a loop
Before, I was executing each run from the command line. This was a
horrible idea. It restarted the entire program which meant that the
entire image dataset; 20k images; had to be reloaded each time. Putting
a loop, after the images are loaded and standardized, but before the
creation and running of the CNN, has sped up execution time
tremendously.